requirements.txt
torch
transformers
pillow
sentencepiece
protobuf